% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tipi.R
\docType{data}
\name{YYMMDD_TYPE}
\alias{YYMMDD_TYPE}
\title{Tipi utilizzati per i dati di temperatura e precipitazione

Questa e' la specifica dei tipi che viene passata a readr::red_delim per la lettura dei dati giornalieri di temperatura e precipitazione}
\format{
An object of class \code{col_spec} of length 2.
}
\usage{
YYMMDD_TYPE
}
\description{
Tipi utilizzati per i dati di temperatura e precipitazione

Questa e' la specifica dei tipi che viene passata a readr::red_delim per la lettura dei dati giornalieri di temperatura e precipitazione
}
\keyword{datasets}
